streamlit_app/requirements.txt
streamlit>=1.32.0
openai>=1.0.0
python-dotenv>=1.0.0
uuid
